requirements.txt
PyHamcrest
requests
pyjq
yapf
pyyaml==3.13
lxml
configparser